fix(rangeSlider): update cursor for disabled state and fix stories
