+++ core/components/atoms/rangeSlider/__stories__/CustomLabels.story.jsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { RangeSlider } from '@/index';
 
 // CSF format story
-export const cutsomLabels = () => {
+export const customLabels = () => {
   const [value, setValue] = React.useState([2, 4]);
 
   const onChange = (newValue) => {
@@ -10,7 +10,7 @@ export const cutsomLabels = () => {
   };
 
   const labelRenderer = (newValue) => {
-    return `${newValue}%`;
+    return <div>${newValue}%</div>;
   };
 
   return (
@@ -22,7 +22,7 @@ export const cutsomLabels = () => {
       value={value}
       onChange={onChange}
       labelRenderer={labelRenderer}
-      className="mt-7"
+      className="my-8"
     />
   );
 };
@@ -42,17 +42,18 @@ const customCode = `() => {
     <RangeSlider
       min={1}
       max={10}
+      value={value}
       stepSize={0.1}
       labelStepSize={1}
-      value={value}
       onChange={onChange}
       labelRenderer={labelRenderer}
+      className="my-8"
     />
   );
 }`;
 
 export default {
-  title: 'Components/RangeSlider/Variants/Cutsom Labels',
+  title: 'Components/RangeSlider/Custom Labels',
   component: RangeSlider,
   parameters: {
     docs: {
