+++ core/components/atoms/rangeSlider/__stories__/Disabled.story.jsx
@@ -2,21 +2,20 @@ import * as React from 'react';
 import { RangeSlider } from '@/index';
 
 // CSF format story
-export const uncontrolledSlider = () => {
+export const disabled = () => {
   return (
     <RangeSlider
-      min={1}
-      max={10}
-      label="Uncontrolled Slider"
+      className="my-8"
+      label="Disabled Slider"
+      disabled={true}
       stepSize={0.1}
       labelStepSize={1}
       defaultValue={[2, 4]}
-      className="mt-2"
     />
   );
 };
 
 export default {
-  title: 'Components/RangeSlider/Variants/Uncontrolled Slider',
+  title: 'Components/RangeSlider/Disabled',
   component: RangeSlider,
 };
