+++ docs/src/pages/components/slider/usage.mdx
@@ -20,7 +20,7 @@ The default slider can be used whenever selection of a single value is required 
 
 Range slider is used if the user wants to select a range instead of a single value.
 
-<Preview name='components-rangeslider-variants-uncontrolled-slider--uncontrolled-slider'/>
+<Preview name='components-rangeslider-types-uncontrolled-slider--uncontrolled-slider'/>
 
 ### Variants
